dts: msm8998-cheeseburger: Setup simplefb using bootloader's framebuffer
With this we can repurpose the bootloader reserved display area normally
used to display a boot animation for simplefb.

This works because the display was already left in an enabled state by
the bootloader itself :^)

If you're planning to copy these you may also need to add the following
properties under /reserved-memory (as well as enable CONFIG_FB_SIMPLE):

  #address-cells = <2>;
  #size-cells = <2>;
  ranges;

NOTE: This also breaks UART as stdout-path is changed.