dts: msm8998-cheeseburger: Add Synaptics RMI4 I2C touchscreen node
As it turns out the downstream Synaptics S3320 driver works via RMI4 and
the mainline "syna,rmi4-i2c" compatible DT bindings do the job of making
the touchscreen on the OnePlus 5 devices work!

Massive thanks to r0551 (@Rossi15) for working on this and figuring it
all out!

I also adapted this to the 5T (dumpling) by changing the digitizer
height and this seems to have been enough to get it working properly
after further testing ^^