dts: msm8998-cheeseburger: Add bindings for Samsung S6E3FA5 panel
Eventually this should turn into a proper commit that enables the
display on the OnePlus 5 to work properly and allow Adreno 540 GPU
acceleration etc. to be implemented :)

Right now though we're still not quite there and the device just crashes
on boot.

For debugging the following defconfig options can be used:

  CONFIG_FTRACE=y
  CONFIG_BOOTTIME_TRACING=y
  CONFIG_FUNCTION_TRACER=y

Device Tree -wise you can add "loglevel=10 drm.debug=15 debug" to the
bootargs in msm8998-cheeseburger.dtsi for increased logging verbosity.