dts: msm8998-cheeseburger: WIP: Add GPIO bindings for the tri-state key
With these testing bindings you can change the state of the tri-state
key and depending on the position it will act like the key 'a' (top),
'b' (middle) or 'c' (bottom).

Sometimes the previous state's key gets activated during transition for
whatever reason which is why I'm keeping this as W.I.P for now.

Downstream resources:

  https://github.com/LineageOS/android_kernel_oneplus_msm8998/blob/lineage-16.0/arch/arm/boot/dts/qcom/cheeseburger.dtsi#L133-L145
  https://github.com/LineageOS/android_kernel_oneplus_msm8998/blob/lineage-16.0/arch/arm/boot/dts/qcom/cheeseburger.dtsi#L74-L101